Z13737 Initial support for AMPS Flume Sink
1. Initial release of an AMPS Flume Sink component that can
be configured as a sink in a Flume message flow to publish
the stream into an AMPS server.

2. Updated AMPSBasicClientFunction to support configuring a
publish store on the AMPS client for reliable publishing.

3. Added config key constants related to the new AMPS sink.

4. Renamed project in Maven POM file and updated versions.

5. Updated README and example config files to have the AMPS
Flume source write its messages into two replicated Flume
memory channels, each of which is read by a separate sink.
The first is read by a standard file_roll sink. The second
is read by the new AMPS Flume sink, which publishes the
stream into another AMPS topic.